+++ mex/common/models/__init__.py
@@ -64,6 +64,7 @@ lists of models, lookups by class name and typing for unions of models.
 from typing import Final, get_args
 
 from mex.common.models.access_platform import (
+    AccessPlatformRuleSet,
     AdditiveAccessPlatform,
     BaseAccessPlatform,
     ExtractedAccessPlatform,
@@ -72,6 +73,7 @@ from mex.common.models.access_platform import (
     SubtractiveAccessPlatform,
 )
 from mex.common.models.activity import (
+    ActivityRuleSet,
     AdditiveActivity,
     BaseActivity,
     ExtractedActivity,
@@ -83,6 +85,7 @@ from mex.common.models.base import BaseModel, GenericFieldInfo
 from mex.common.models.contact_point import (
     AdditiveContactPoint,
     BaseContactPoint,
+    ContactPointRuleSet,
     ExtractedContactPoint,
     MergedContactPoint,
     PreventiveContactPoint,
@@ -91,6 +94,7 @@ from mex.common.models.contact_point import (
 from mex.common.models.distribution import (
     AdditiveDistribution,
     BaseDistribution,
+    DistributionRuleSet,
     ExtractedDistribution,
     MergedDistribution,
     PreventiveDistribution,
@@ -110,6 +114,7 @@ from mex.common.models.organization import (
     BaseOrganization,
     ExtractedOrganization,
     MergedOrganization,
+    OrganizationRuleSet,
     PreventiveOrganization,
     SubtractiveOrganization,
 )
@@ -118,6 +123,7 @@ from mex.common.models.organizational_unit import (
     BaseOrganizationalUnit,
     ExtractedOrganizationalUnit,
     MergedOrganizationalUnit,
+    OrganizationalUnitRuleSet,
     PreventiveOrganizationalUnit,
     SubtractiveOrganizationalUnit,
 )
@@ -126,6 +132,7 @@ from mex.common.models.person import (
     BasePerson,
     ExtractedPerson,
     MergedPerson,
+    PersonRuleSet,
     PreventivePerson,
     SubtractivePerson,
 )
@@ -135,6 +142,7 @@ from mex.common.models.primary_source import (
     ExtractedPrimarySource,
     MergedPrimarySource,
     PreventivePrimarySource,
+    PrimarySourceRuleSet,
     SubtractivePrimarySource,
 )
 from mex.common.models.resource import (
@@ -143,6 +151,7 @@ from mex.common.models.resource import (
     ExtractedResource,
     MergedResource,
     PreventiveResource,
+    ResourceRuleSet,
     SubtractiveResource,
 )
 from mex.common.models.rules import AdditiveRule, PreventiveRule, SubtractiveRule
@@ -153,6 +162,7 @@ from mex.common.models.variable import (
     MergedVariable,
     PreventiveVariable,
     SubtractiveVariable,
+    VariableRuleSet,
 )
 from mex.common.models.variable_group import (
     AdditiveVariableGroup,
@@ -161,6 +171,7 @@ from mex.common.models.variable_group import (
     MergedVariableGroup,
     PreventiveVariableGroup,
     SubtractiveVariableGroup,
+    VariableGroupRuleSet,
 )
 
 __all__ = (
@@ -369,6 +380,26 @@ AnyRuleModel = AnyAdditiveModel | AnySubtractiveModel | AnyPreventiveModel
 RULE_MODEL_CLASSES: Final[list[type[AnyRuleModel]]] = list(get_args(AnyRuleModel))
 RULE_MODEL_CLASSES_BY_NAME: Final[dict[str, type[AnyRuleModel]]] = {
     cls.__name__: cls for cls in RULE_MODEL_CLASSES
+}
+
+AnyRuleSetModel = (
+    AccessPlatformRuleSet
+    | ActivityRuleSet
+    | ContactPointRuleSet
+    | DistributionRuleSet
+    | OrganizationRuleSet
+    | OrganizationalUnitRuleSet
+    | PersonRuleSet
+    | PrimarySourceRuleSet
+    | ResourceRuleSet
+    | VariableRuleSet
+    | VariableGroupRuleSet
+)
+RULE_SET_MODEL_CLASSES: Final[list[type[AnyRuleSetModel]]] = list(
+    get_args(AnyRuleSetModel)
+)
+RULE_SET_MODEL_CLASSES_BY_NAME: Final[dict[str, type[AnyRuleSetModel]]] = {
+    cls.__name__: cls for cls in RULE_SET_MODEL_CLASSES
 }
 
 FILTER_MODEL_BY_EXTRACTED_CLASS_NAME = {