+++ mex/common/models/person.py
@@ -7,7 +7,12 @@ from pydantic import Field, computed_field
 from mex.common.models.base import BaseModel
 from mex.common.models.extracted_data import ExtractedData
 from mex.common.models.merged_item import MergedItem
-from mex.common.models.rules import AdditiveRule, PreventiveRule, SubtractiveRule
+from mex.common.models.rules import (
+    AdditiveRule,
+    PreventiveRule,
+    RuleSet,
+    SubtractiveRule,
+)
 from mex.common.types import (
     Email,
     ExtractedPersonIdentifier,
@@ -135,3 +140,14 @@ class PreventivePerson(_Stem, PreventiveRule):
     isniId: list[MergedPrimarySourceIdentifier] = []
     memberOf: list[MergedPrimarySourceIdentifier] = []
     orcidId: list[MergedPrimarySourceIdentifier] = []
+
+
+class PersonRuleSet(_Stem, RuleSet):
+    """Set of rules to edit a person item."""
+
+    entityType: Annotated[
+        Literal["PersonRuleSet"], Field(alias="$type", frozen=True)
+    ] = "PersonRuleSet"
+    additive: AdditivePerson
+    subtractive: SubtractivePerson
+    preventive: PreventivePerson