+++ mex/common/models/variable_group.py
@@ -7,7 +7,12 @@ from pydantic import Field, computed_field
 from mex.common.models.base import BaseModel
 from mex.common.models.extracted_data import ExtractedData
 from mex.common.models.merged_item import MergedItem
-from mex.common.models.rules import AdditiveRule, PreventiveRule, SubtractiveRule
+from mex.common.models.rules import (
+    AdditiveRule,
+    PreventiveRule,
+    RuleSet,
+    SubtractiveRule,
+)
 from mex.common.types import (
     ExtractedVariableGroupIdentifier,
     MergedPrimarySourceIdentifier,
@@ -90,3 +95,14 @@ class PreventiveVariableGroup(_Stem, PreventiveRule):
     ] = "PreventiveVariableGroup"
     containedBy: list[MergedPrimarySourceIdentifier] = []
     label: list[MergedPrimarySourceIdentifier] = []
+
+
+class VariableGroupRuleSet(_Stem, RuleSet):
+    """Set of rules to edit a variable group item."""
+
+    entityType: Annotated[
+        Literal["VariableGroupRuleSet"], Field(alias="$type", frozen=True)
+    ] = "VariableGroupRuleSet"
+    additive: AdditiveVariableGroup
+    subtractive: SubtractiveVariableGroup
+    preventive: PreventiveVariableGroup