+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - convenience function `get_merged_organization_id_by_query_with_extract_transform_and_load`
   for getting the stableTargetId of an organization, while transforming and loading the
   organization using the provided load function
+- models for rule sets along with typing and lookups
 
 ### Changes